# docs/swagger.yaml
definitions:
...
  dto.Order:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/dto.OrderItem'
        type: array
      status:
        type: string
      store_id:
        type: integer
      total_price:
        type: number
      user_id:
        type: integer
    type: object
  dto.OrderItem:
    properties:
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  dto.PaginatedResponse:
    properties:
      data: {}
      total:
        type: integer
    type: object
...
  errorx.AppError:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        $ref: '#/definitions/errorx.ErrorCode'
      message:
        type: string
    type: object
  errorx.ErrorCode:
    enum:
    - UNAUTHORIZED
    - FORBIDDEN
    - VALIDATION_ERROR
    - NOT_FOUND
    - INTERNAL_ERROR
    - BAD_REQUEST
    - CONFLICT
    type: string
    x-enum-varnames:
    - ErrUnauthorized
    - ErrForbidden
    - ErrValidation
    - ErrNotFound
    - ErrInternal
    - ErrBadRequest
    - ErrConflict
  response.BaseResponse:
    properties:
      data: {}
      error: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /order/create:
    post:
      description: Creates a new order based on the user's current cart items for
        a specific store
      parameters:
      - description: Store ID
        in: query
        name: store_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Create order from cart
      tags:
      - Order
  /orders/:order_id:
    get:
      description: Retrieve details of a specific order by ID
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Order'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Get order
      tags:
      - Order
  /payment/verify:
    post:
      consumes:
      - application/json
      description: Verify manual payment order and credit wallet
      parameters:
      - description: Payment Details
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Verify payment (manual)
      tags:
      - Payment
...
  /store/:store_id/orders:
    get:
      description: Retrieve a paginated list of all orders for a given store
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: List orders
      tags:
      - Order
  /store/:store_id/orders/:order_id/status:
    put:
      description: Update the status of an existing order
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: New Status
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Update order status
      tags:
      - Order
...
  /store/:store_id/wallet:
    get:
      description: Retrieve the wallet balance for a given store
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: number
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Get wallet balance
      tags:
      - Wallet
...
  /user/orders:
    get:
      description: Retrieve a paginated list of all orders belonging to the authenticated
        user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: List user orders
      tags:
      - Order
  /user/orders/:order_id/cancel:
    put:
      description: Cancels a pending order and restores product stock
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Cancel order
      tags:
      - Order
swagger: "2.0"